+++ ChangeLog.md
@@ -45,7 +45,7 @@
 * Update some config.
 
 ### 2021-04-27 Version 0.2.7
-* Throw TeaException will not retry.
+* Throw DaraException will not retry.
 * Improve instance check of stream.
 
 ### 2021-03-17 Version 0.2.6
@@ -71,7 +71,7 @@
 * Drop support for python3.4.
 * Drop support for python3.5.
 * Added support for python3.9.
-* Improve TeaException error message.
+* Improve DaraException error message.
 * Support async do action.
 
 ### 2020-10-30 Version 0.1.5